DataMovementEventPayloadProto contains containerId.
Tez spill records are now kept in memory (when using tez_shuffle),
so InputHosts needs to distinguish between not just hosts but also ContainerWorkers.
Otherwise InputAttemptIdentifiers from different ContainerWorkers might be added
to the same InputHost (e.g., when the first ContainerWorker is killed and the second
ContainerWorker is created on the same host).